# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "description": "Public. Prices and request limits of active plans.",
        "operationId": "PlansController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit or plan quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 429,
                  "code": "RATE_LIMITED",
                  "message": "Too many requests, please slow down",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "summary": "List available plans",
        "tags": ["Subscriptions"]
      }
    },
...
    "/api/v1/admin/subscriptions": {
      "get": {
        "description": "Filter by plan code and status. Newest first.",
        "operationId": "AdminSubscriptionsController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "planCode",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 32,
              "example": "premium",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ACTIVE",
              "type": "string",
              "enum": ["ACTIVE", "CANCELED", "EXPIRED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminSubscriptionResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the request is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_FAILED",
                  "message": "Request validation failed",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "FORBIDDEN",
                  "message": "Forbidden",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List subscriptions",
        "tags": ["Admin · Subscriptions"]
      }
    },
    "/api/v1/admin/users/{id}/subscription": {
      "put": {
        "description": "Same semantics as a self-service plan change: the active subscription is cancelled and a new one starts.",
        "operationId": "AdminSubscriptionsController_assignPlan_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the request is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_FAILED",
                  "message": "Request validation failed",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "FORBIDDEN",
                  "message": "Forbidden",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or not owned by the caller)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 404,
                  "code": "PLAN_NOT_FOUND",
                  "message": "Plan \"gold\" does not exist",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state of a resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 409,
                  "code": "PLAN_ALREADY_ACTIVE",
                  "message": "You are already on the Premium plan",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Assign a plan to a user",
        "tags": ["Admin · Subscriptions"]
      }
    },
    "/api/v1/admin/plans": {
      "get": {
        "operationId": "AdminSubscriptionsController_listPlans_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminPlanResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "FORBIDDEN",
                  "message": "Forbidden",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all plans, including inactive ones",
        "tags": ["Admin · Subscriptions"]
      }
    },
    "/api/v1/admin/plans/{id}": {
      "patch": {
        "description": "Changing limits applies immediately to every subscriber of the plan.",
        "operationId": "AdminSubscriptionsController_updatePlan_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the request is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_FAILED",
                  "message": "Request validation failed",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token, or revoked session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "UNAUTHORIZED",
                  "message": "Unauthorized",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "FORBIDDEN",
                  "message": "Forbidden",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or not owned by the caller)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 404,
                  "code": "PLAN_NOT_FOUND",
                  "message": "Plan not found",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          },
          "422": {
            "description": "Request is valid but violates a business rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "example": {
                  "statusCode": 422,
                  "code": "PLAN_IS_DEFAULT",
                  "message": "The default plan for new accounts cannot be deactivated",
                  "details": null,
                  "path": "/api/v1/…",
                  "timestamp": "2026-09-26T10:00:00.000Z",
                  "requestId": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Edit a plan",
        "tags": ["Admin · Subscriptions"]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "REST API for the EchoGPT multi-AI chat Chrome extension.\n\n**Authentication:** call `POST /auth/login`, then send `Authorization: Bearer <accessToken>`.\nRefresh with `POST /auth/refresh` before the access token expires.\n\n**Errors** always use the `ErrorResponseDto` envelope with a stable `code`.\n**Tracing:** send `x-request-id` to correlate logs; it is echoed back on every response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable error code"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Extra context, e.g. field-level validation errors",
            "example": null
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource/0192f0c4-8a3e-7c1e-9d2b-5b1f3a9c2e10"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "4f0c8a2e-1b3d-4c5e-9f6a-7b8c9d0e1f2a"
          }
        },
        "required": ["statusCode", "code", "message", "path", "timestamp", "requestId"]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "premium"
          },
          "name": {
            "type": "string",
            "example": "Premium"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Higher daily allowance for power users"
          },
          "priceCents": {
            "type": "number",
            "description": "Price in the smallest currency unit",
            "example": 999
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "requestLimit": {
            "type": "number",
            "description": "Requests allowed per period",
            "example": 500
          },
          "limitPeriod": {
            "enum": ["DAILY", "MONTHLY"],
            "type": "string",
            "example": "DAILY"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "priceCents",
          "currency",
          "requestLimit",
          "limitPeriod"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": ["ACTIVE", "CANCELED", "EXPIRED"],
            "type": "string",
            "example": "ACTIVE"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanResponseDto"
          },
          "currentPeriodStart": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "canceledAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "plan",
          "currentPeriodStart",
          "currentPeriodEnd",
          "canceledAt",
          "createdAt"
        ]
      },
...
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "pattern": "^[a-z0-9_-]+$",
            "example": "premium",
            "description": "Code of the target plan (see GET /plans)"
          }
        },
        "required": ["planCode"]
      },
      "SubscriptionUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Doe"
          }
        },
        "required": ["id", "email", "fullName"]
      },
      "AdminSubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": ["ACTIVE", "CANCELED", "EXPIRED"],
            "type": "string",
            "example": "ACTIVE"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanResponseDto"
          },
          "currentPeriodStart": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-26T10:00:00.000Z"
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "canceledAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SubscriptionUserDto"
          }
        },
        "required": [
          "id",
          "status",
          "plan",
          "currentPeriodStart",
          "currentPeriodEnd",
          "canceledAt",
          "createdAt",
          "user"
        ]
      },
      "AdminPlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "premium"
          },
          "name": {
            "type": "string",
            "example": "Premium"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Higher daily allowance for power users"
          },
          "priceCents": {
            "type": "number",
            "description": "Price in the smallest currency unit",
            "example": 999
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "requestLimit": {
            "type": "number",
            "description": "Requests allowed per period",
            "example": 500
          },
          "limitPeriod": {
            "enum": ["DAILY", "MONTHLY"],
            "type": "string",
            "example": "DAILY"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "priceCents",
          "currency",
          "requestLimit",
          "limitPeriod",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "example": "Premium"
          },
          "description": {
            "type": "string",
            "minLength": 0,
            "maxLength": 255,
            "example": "Higher daily allowance for power users"
          },
          "priceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000000,
            "example": 1499,
            "description": "Price in cents"
          },
          "requestLimit": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000000,
            "example": 1000,
            "description": "Requests per period; applies immediately to all subscribers"
          },
          "limitPeriod": {
            "enum": ["DAILY", "MONTHLY"],
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "description": "Inactive plans are hidden from GET /plans and cannot be chosen"
          }
        }
      },
...
    }
  }
}